Bug fix. Moved user edit PHP to separate file (userEdits.php)
